refactor: re-organize api directory structure
This commit simply splits the `index.js` file in many modules organized
by the api endpoints.

The `config` key in the api is re-mapped to `workspaces`, since it was
only concerning operations related to the workspaces configuration.
